requirements.txt
Flask==2.3.0
Flask-Cors==4.0.0
Pillow==10.0.0
numpy==1.24.3
tensorflow==2.15.0
Werkzeug==2.3.0
Jinja2==3.1.2
gunicorn==21.2.0